% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-matrix-manipulation.R
\name{averageReplicate}
\alias{averageReplicate}
\title{Average Replicates Gene Wise}
\usage{
averageReplicate(matrix)
}
\arguments{
\item{matrix}{with columns names from which infer the factors}
}
\value{
averaged matrix
}
\description{
Average Replicates Gene Wise
}
