% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-objects.R
\name{defineCellExpression}
\alias{defineCellExpression}
\title{Define Cell Expression}
\usage{
defineCellExpression(day, binCell, clusterCell, measurableGenes = NULL)
}
\arguments{
\item{day}{the day}

\item{binCell}{binary cluster representation}

\item{clusterCell}{the gene organized in clusters}

\item{measurableGenes}{the total measurable genes}
}
\value{
list with activeAtDay, non_avtiveAtDay and constitutive
}
\description{
Formely in findAWay R package
}
