% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova-implementations.R
\name{edgeR.allCmpOnTimeWithRUV}
\alias{edgeR.allCmpOnTimeWithRUV}
\title{ANOVA like analysis (edgeR)}
\usage{
edgeR.allCmpOnTimeWithRUV(
  eset,
  unvariant,
  k = 1,
  formula = "~days + W_1",
  varName = "days"
)
}
\arguments{
\item{eset}{expression set}

\item{unvariant}{set of unvariant genes}

\item{k}{number of covariate}

\item{formula}{a string that can be converted in formula}

\item{varName}{the name of the Pdata variable to use for the anova}
}
\description{
ANOVA like analysis (edgeR)
}
