% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova-implementations.R
\name{edgeR.anovaOnTime}
\alias{edgeR.anovaOnTime}
\title{ANOVA like analysis (edgeR)}
\usage{
edgeR.anovaOnTime(eset, k = 1, formula = "~days", varName = "days")
}
\arguments{
\item{eset}{expression set}

\item{k}{number of covariate}

\item{formula}{a string that can be converted in formula}

\item{varName}{the name of the Pdata variable to use for the anova}
}
\description{
ANOVA like analysis (edgeR)
}
