% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsbk-gene-set-handling.R
\name{convertGskb2Term2Gene}
\alias{convertGskb2Term2Gene}
\alias{convertGskb2Term2Name}
\alias{summaryBarplot}
\title{Convert to Gene}
\usage{
convertGskb2Term2Gene(list)

convertGskb2Term2Name(list)

summaryBarplot(enrichmentList, single = T, title = "cluster",
  ncol = 3, font.size = 5, showCategory = 100)
}
\arguments{
\item{list}{a list}

\item{enrichmentList}{a list}

\item{single}{whether to plot single page plots or cumulative plots}

\item{title}{suffix names}

\item{ncol}{the column}

\item{font.size}{the font size}

\item{showCategory}{max number of category to show}
}
\value{
term2gene

a data.frame with term2name

a data.frame with term2name
}
\description{
Convert to Gene

Convert to Name

create summarizing barplots from enrichment lists
}
