% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-objects.R
\name{extractDegsPvaluesWithLFC}
\alias{extractDegsPvaluesWithLFC}
\alias{extractDegsWithLFC}
\alias{getDEGsPvaluesFromObj}
\alias{getActiveAtDaysFromObj}
\title{Collect DEGs Pvalues with LogFold Change from RData.}
\usage{
extractDegsPvaluesWithLFC(esetRUV, pADJ.thr = 0.05, logfcthr = 1)

extractDegsWithLFC(esetRUV, pADJ.thr = 0.05, logfcthr = 1)

getDEGsPvaluesFromObj(obj, adjusted = T)

getActiveAtDaysFromObj(obj)
}
\arguments{
\item{esetRUV}{an expression set}

\item{pADJ.thr}{the threshold for adjusted pvalue, default 0.05}

\item{logfcthr}{the absolute threshold value for lfc, default 1}

\item{obj}{expression set}

\item{adjusted}{logical, if true retrieve adjusted pvalues}
}
\value{
a three column matix with gene, pvalue and lfc

a three column matix with gene, pvalue and lfc

pvalue vector

a list of genes active at Days.
}
\description{
Collect DEGs Pvalues with LogFold Change from RData.

Collect DEGs LogFold Change from RData.

Load and collect Active at Days from Obj.

Load and collect Active at Days from Obj.
}
