% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.RData.R
\name{getNormExpressionFromRData}
\alias{getNormExpressionFromRData}
\alias{getRpkmExpressionFromRData}
\alias{extractDegsPvaluesWithLFCFromRData}
\alias{getDEGsPvaluesFromRData}
\alias{getActiveAtDaysFromRData}
\title{Load and extract normalized expression from RData.}
\usage{
getNormExpressionFromRData(RData, objname)

getRpkmExpressionFromRData(RData, objname)

extractDegsPvaluesWithLFCFromRData(RData, objname, pADJ.thr = 0.05,
  logfcthr = 1.5)

getDEGsPvaluesFromRData(RData, objname)

getActiveAtDaysFromRData(RData, objname)
}
\arguments{
\item{RData}{filnename with RData}

\item{objname}{the name of the loaded object}

\item{pADJ.thr}{the threshold for adjusted pvalue, default 0.05}

\item{logfcthr}{the absolute threshold value for lfc, default 1.5}
}
\value{
normalized gene expression matrix

rpkm expression matrix

a three column matix with gene, pvalue and lfc

a list of genes Expressed at a given Day.

a list of genes active at a given Day.
}
\description{
Load and extract normalized expression from RData.

Load and collect RPKM from RData.

Load and collect DEGs Pvalues with LogFold Change from RData.

Load and collect Active at Days from RData.

Load and collect Active Genes at Days from RData.
}
