% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-fold-change-handling.R
\name{anyCmpSigByLogFC}
\alias{anyCmpSigByLogFC}
\alias{anyLogFC}
\title{Extract the contrasts with a given log fold change (lfc) from anova-like (edgeR) lfc table}
\usage{
anyCmpSigByLogFC(x, cmp, lfcThr = 2, baseCellCmp = "ec")

anyLogFC(x, lfcThr = 2)
}
\arguments{
\item{x}{a vector of logFoldChanges}

\item{cmp}{the vector of contrasts, typically colnames of lfc table}

\item{lfcThr}{the lfc threshold (in absolute value)}

\item{baseCellCmp}{the basal cell used in comparison}
}
\value{
a data.frame with
   \item{gene}{the gene}
   \item{num}{the numerator of the contrast}
   \item{den}{the denominator of the contrast}
   \item{winner}{the winner of the contrast}
   \item{looser}{the loosers of the contrast}

TRUE / FALSE
}
\description{
Extract the contrasts with a given log fold change (lfc) from anova-like (edgeR) lfc table

Extract All Genes with given log fold change (lfc) from anova-like lfc table
}
