% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt-list-into-tables.R
\name{meltTablesWithListNames}
\alias{meltTablesWithListNames}
\alias{meltTablesWithLoosers}
\alias{meltListOfVectors}
\title{Melt list of tables}
\usage{
meltTablesWithListNames(list)

meltTablesWithLoosers(list)

meltListOfVectors(list)
}
\arguments{
\item{list}{a list of tables}
}
\value{
a data frame with
   \item{gene}{the gene name}
   \item{cell}{the cell}
   \item{cpmWin}{the number of wins it collected}

a data frame with
   \item{gene}{the gene name}
   \item{cell}{the cell who win}
   \item{loosers}{the semi-colon separated list of loosers}

a data frame with list content and the name of the element in the column
}
\description{
Tables are the output of tables

Tables are the output of tables
}
