% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{mergeColumns}
\alias{mergeColumns}
\title{Merge columns}
\usage{
mergeColumns(li, col = "p.adjust", cutOff = 0.05)
}
\arguments{
\item{li}{a list objects dataframe like}

\item{col}{the name of the column}

\item{cutOff}{set to NA if lower than cutOff}
}
\value{
a matrix like
}
\description{
Merge columns
}
