% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-profiles-hijack.R
\name{my.compareCluster}
\alias{my.compareCluster}
\title{Perform compare cluster}
\usage{
my.compareCluster(geneClusters, go_bp_full_data, data = "", ...)
}
\arguments{
\item{geneClusters}{gene}

\item{go_bp_full_data}{go biological processed}

\item{data}{NULL}

\item{\dots}{other argument}
}
\description{
Perform compare cluster
}
