% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-profiles-hijack.R
\name{my.enricher}
\alias{my.enricher}
\title{Perform enricher with a slight change}
\usage{
my.enricher(gene, go_bp_full_data, pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe, qvalueCutoff = 0.2, minGSSize = 10,
  maxGSSize = 500, readable = FALSE, pool = FALSE, ont = "BP",
  keyType = "SYMBOL", OrgDb = "org.Mm.eg.db")
}
\arguments{
\item{gene}{gene}

\item{go_bp_full_data}{go biological processed}

\item{pvalueCutoff}{pvalue cut off}

\item{pAdjustMethod}{BH}

\item{universe}{the full list of genes}

\item{qvalueCutoff}{0.2}

\item{minGSSize}{10}

\item{maxGSSize}{500}

\item{readable}{FALSE}

\item{pool}{FALSE}

\item{ont}{BP}

\item{keyType}{SYMBOL}

\item{OrgDb}{org.Mm.eg.db}
}
\description{
Perform enricher with a slight change
}
