% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paracrine-autocrine-analysis.R
\name{getCellExpressionSignature}
\alias{getCellExpressionSignature}
\alias{extractGenesBehavious}
\alias{extractLigandReceptorBehaviours}
\alias{assignColorCode}
\alias{assignCellCode}
\alias{resolveWarnings}
\alias{removeElement}
\alias{wiseMergeTimePoints}
\alias{paraWiseMergeTimePoints}
\alias{paraWiseMultiMergeTimePoints}
\alias{renderHTMLvideo}
\alias{renderHTMLvideo2.0}
\title{get Cell Expression Signature}
\usage{
getCellExpressionSignature(
  RData = "rpkm-Rdatas/all-wt.RData",
  objname = "all_wt",
  cols = 1:4
)

extractGenesBehavious(activeAtDays, activeAtDaysTot, constitutiveWide)

extractLigandReceptorBehaviours(genesBehaviours, db)

assignColorCode(ligandReceptorBehaviours, paletteName = "Greens")

assignCellCode(ligandReceptorBehaviours, cell)

resolveWarnings(lrCols, ligands, receptors)

removeElement(namedVector, nms)

wiseMergeTimePoints(ligandReceptorBehaviours)

paraWiseMergeTimePoints(
  ligandReceptorBehavioursCell,
  ligCell = "fap",
  recCell = "ec"
)

paraWiseMultiMergeTimePoints(
  ligandReceptorBehavioursCell,
  ligCells = c("ec", "fap", "mp", "inf", "per"),
  recCell = "ec",
  removeAutocrine = TRUE,
  useConstitutiveLigands = FALSE
)

renderHTMLvideo(
  ligandsReceptorCumulative,
  receptorGraph,
  lrCols,
  lrCell = NULL,
  cell,
  condition
)

renderHTMLvideo2.0(ligandsReceptorCumulative, receptorGraph, cell, condition)
}
\arguments{
\item{RData}{an data RData}

\item{objname}{the loaded object name}

\item{cols}{columns to use}

\item{activeAtDays}{list of gene active at days}

\item{activeAtDaysTot}{list of gene active in total tissue}

\item{constitutiveWide}{constitutive active genes}

\item{genesBehaviours}{list of gene active at days from extractGenesBehavious}

\item{db}{the ligand receptor database}

\item{ligandReceptorBehaviours}{list of ligand and receptor ativation}

\item{paletteName}{the name of the palette to use.}

\item{cell}{cell name}

\item{lrCols}{dictionary of ligand receptor colors}

\item{ligands}{vector of ligands}

\item{receptors}{vector of receptors}

\item{namedVector}{named vector}

\item{nms}{names}

\item{ligandReceptorBehavioursCell}{list of ligand enad receptor ativation}

\item{ligCell}{cell that produce the ligands}

\item{recCell}{cell that produce the receptors}

\item{ligCells}{cells that produce the ligands}

\item{removeAutocrine}{if TRUE remove aoutocrine loops}

\item{useConstitutiveLigands}{if TRUE use constitutive ligands}

\item{ligandsReceptorCumulative}{list of ligand receptors per day}

\item{receptorGraph}{the graph receptor}

\item{lrCell}{dictionary of ligand receptor cells}

\item{condition}{cell condition}
}
\description{
get Cell Expression Signature

Define gene behaviour in the cell

Define ligand receptor behaviour in the cell

Assign color to the nodes

Assign color to the nodes

Assign color to the nodes

Assign color to the nodes

Merge time points wisely

Merge time points wisely

Merge time points wisely

Render Html Video

Render the video
}
