% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runWithParametersTotalRpkm}
\alias{runWithParametersTotalRpkm}
\title{Run in total muscle [DEPRECATED]}
\usage{
runWithParametersTotalRpkm(dataFile, annotationFile, cell, condition,
  runWithK = 1, runWithFormula = "~days + W_1", varName = "days",
  pADJ.thr = 0.05, logfcthr = 1.5, rcFilter = 100, atLeastIn = 2,
  rpkmMean = 3, clusterK = 9, gene_length = "gene_length.txt")
}
\arguments{
\item{dataFile}{the name of the file that contains the expression dataset}

\item{annotationFile}{the name of the file that contains annotations}

\item{cell}{the cell sample name to pick}

\item{condition}{the condition}

\item{runWithK}{run RUV using k covariates}

\item{runWithFormula}{a string coercible with as.formula}

\item{varName}{the variable to used to evaluate unwanted variance}

\item{pADJ.thr}{the thresholf for pvalue adjusted}

\item{logfcthr}{absolute log fold change}

\item{rcFilter}{filter expression for genes minimal rowCounts in atLeastIn samples}

\item{atLeastIn}{filter expression for genes minimal rowCounts in atLeastIn samples}

\item{rpkmMean}{filter expression for genes with an average rpkm of 3}

\item{clusterK}{try clusterize according to k clusters}

\item{gene_length}{filename of file containg gene lengths}
}
\description{
Run in total muscle [DEPRECATED]
}
