% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV-handling.R
\name{createSet}
\alias{createSet}
\alias{createSeqExpressionSet}
\alias{computeEmpiricalUnvariants}
\alias{computeResiduals}
\alias{filterExps}
\alias{filterExpression}
\alias{filterExpressionByRpkmMean}
\alias{computeRPKM}
\title{Create an expression set using EDASeq function}
\usage{
createSet(obj)

createSeqExpressionSet(expression, annotation)

computeEmpiricalUnvariants(set, number = 5000, varName = "days")

computeResiduals(set)

filterExps(cell, condition, expr, annotations, rcFilter = NULL)

filterExpression(
  cell,
  condition,
  expr,
  annotations,
  rcFilter = NULL,
  atLeastIn = 1,
  unvariantNumber = 5000,
  varName = "days"
)

filterExpressionByRpkmMean(
  cell,
  condition,
  expr,
  annotations,
  rcFilter = NULL,
  atLeastIn = 1,
  rpkmMean = NULL,
  unvariantNumber = 5000,
  varName = "days",
  gene_length = "gene_length.txt"
)

computeRPKM(
  eset,
  gene_length = "../m38.p5-ensembl-91-renge-length/gene_length.txt"
)
}
\arguments{
\item{obj}{a list with $e expression and $ann annotation}

\item{expression}{expression data}

\item{annotation}{annotation data}

\item{set}{an expression set}

\item{number}{the number of desired invariant genes}

\item{varName}{the variable that set up the contrast to compute invariants; default "days"}

\item{cell}{string for cell name (can be ALL or a vector)}

\item{condition}{string for cell condition (can be a vector with more than one)}

\item{expr}{data matrix}

\item{annotations}{annotation data referred to the matrix}

\item{rcFilter}{minimal read count accepted in atLeastIn samples}

\item{atLeastIn}{default 1 sample}

\item{unvariantNumber}{number of invariant genes; default 5000}

\item{rpkmMean}{minimal RPKM row mean accepted to keep the row; NULL}

\item{gene_length}{filename of file containg gene lengths}

\item{eset}{expression set}
}
\value{
expression set (eset)

expression set (eset)

a vector of invariant gene

a vector residuals

a list with filtered expression, filtered annotation

a filtered eset and other details

a filtered eset and other details
}
\description{
Create an expression set using EDASeq function

Create an expression set using EDASeq function

Compute empirically invariant genes

Compute residuals

Filter expression

Filter expression

Filter expression using RPKM

Compute RPKM from eset
}
