% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-go-terms.R
\name{sigTable_filter_by_clusters}
\alias{sigTable_filter_by_clusters}
\title{Filter significance table pickeing 1 term for cluster}
\usage{
sigTable_filter_by_clusters(sigTable, clusters, lowerCut = 2,
  upperCut = Inf)
}
\arguments{
\item{sigTable}{the table with significant results}

\item{clusters}{the GO organized in clusters}

\item{lowerCut}{excludes the GO with fewer than lowerCut significant times}

\item{upperCut}{excludes the GO with more than upperCut significant times}
}
\description{
Filter significance table pickeing 1 term for cluster
}
