% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-go-terms.R
\name{sigTable_filter_by_clusters_map}
\alias{sigTable_filter_by_clusters_map}
\title{Filter significance table picking 1 term for cluster_map}
\usage{
sigTable_filter_by_clusters_map(
  sigTable,
  cluster_map,
  lowerCut = 2,
  upperCut = Inf
)
}
\arguments{
\item{sigTable}{the table with significant results}

\item{cluster_map}{the GO organized in clusters}

\item{lowerCut}{excludes the GO with fewer than lowerCut significant times}

\item{upperCut}{excludes the GO with more than upperCut significant times}
}
\description{
Filter significance table picking 1 term for cluster_map
}
